# Conflicts:
#	src/main/java/com/minecolonies/core/entity/citizen/citizenhandlers/CitizenDiseaseHandler.java
#	src/main/resources/assets/minecolonies/lang/manual_en_us.json

+++ src/main/java/com/minecolonies/api/colony/jobs/IJob.java
@@ -198,7 +198,7 @@ public interface IJob<AI extends ITickingStateAI> extends INBTSerializable<Compo
      *
      * @return the modifier of the job.
      */
-    int getDiseaseModifier();
+    double getDiseaseModifier();
 
     /**
      * When job removed (death of citizen or job change).
