+++ src/main/java/com/minecolonies/core/colony/jobs/JobComposter.java
@@ -40,7 +40,7 @@ public class JobComposter extends AbstractJob<EntityAIWorkComposter, JobComposte
     }
 
     @Override
-    public int getDiseaseModifier()
+    public double getDiseaseModifier()
     {
         final int skill = getCitizen().getCitizenSkillHandler().getLevel(getCitizen().getWorkBuilding().getModule(BuildingModules.COMPOSTER_WORK).getPrimarySkill());
         return (int) ((100 - skill)/25.0);
